Add files via upload
I am a Chinese,so my English is not so good.

This is a RPGmodpack.And I think as a modpack it is good.

You can find The big bulidings,and the beautiful biomes.

Everything is in the modpack.

Modpack needs 6GB or 8GBRAM,at least 4GBRAM.
Please use the java11 or java14 to start game.
          Key Features:

Many Quests!

All new biomes, cave and cavern generation!

Revamped Nether and End!

Randomized Dungeons Generating All Throughout The World!

Brand new UI!

The breathing bar is hidden. You need to estimate for yourself how much time you have left!

And more and more!

          Summary:

Can I add/remove certain mods?

Yes,of course.

You can download this modpack and change it as you like. For example:

Change the settings of the modpack, add/remove mods, and even reprint.As long as you can't decompose it.

Have fun!